Revert upstream change which marks all WiFi connections as user connections by default
In Debian we are shipping a custom patch which implements a slightly
different behaviour which was broken by this upstream change.
We check whether the user has the
org.freedesktop.NetworkManager.settings.modify.system permission and
only creates a user connection if the permission check fails. Otherwise
we create a system connection.

Thanks: Dean Serenevy <dean@serenevy.net>
Closes: #916187